Create booker.html
This the html/javascript/bootstrap main program. For version control reasons, the telephone booker aspect of this is best implemented by embedding this in a php file so that your telephone bookers run booker.php. For version-control and security reasons, the booker management aspect (booker.html?mode=viewer&ver=xxx) is best implemented by front-ending the system with a login that calls the system after checking credentials, setting the version parameter as it does so. This signon element is left as an exercise, as is the mechanism for policing it in the php database connection function (likewise not provided, but I'm sure you can work this one out). if you're not bothered by these issues, just run booker.html