Update login_engine.html
Reconsideration of the design notes at the head of the document. It was initially believed that the fact that session codes cannot be shared between subdomains were another reason for using the "trusted user" security design adopted here.  It was thought that the only way of proceeding conventionally would have been to abandon the code-sharing arrangements adopted throughout the design.  Recent work elsewhere has demonstrated that this is not the case. However, the "trusted user" design still seems appropriate in the circumstances.